Spike moving finalization to executiton engine
