+++ subprojects/core/src/main/java/org/gradle/api/internal/tasks/properties/DefaultFinalizingValidatingProperty.java
@@ -16,6 +16,8 @@
 
 package org.gradle.api.internal.tasks.properties;
 
+import org.gradle.internal.execution.LifecycleAwareValue;
+
 public class DefaultFinalizingValidatingProperty extends AbstractValidatingProperty {
     private final PropertyValue value;
     private LifecycleAwareValue lifecycleAware;
