+++ subprojects/core/src/main/java/org/gradle/api/internal/tasks/properties/TaskProperties.java
@@ -21,6 +21,7 @@ import org.gradle.api.NonNullApi;
 import org.gradle.api.file.FileCollection;
 import org.gradle.api.internal.TaskInternal;
 import org.gradle.api.internal.tasks.TaskValidationContext;
+import org.gradle.internal.execution.LifecycleAwareValue;
 
 import java.util.Map;
 import java.util.function.Supplier;
