+++ subprojects/execution/src/main/java/org/gradle/internal/execution/steps/FinalizePropertiesStep.java
@@ -1,0 +1,50 @@
+/*
+ * Copyright 2020 the original author or authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.gradle.internal.execution.steps;
+
+import org.gradle.internal.execution.Context;
+import org.gradle.internal.execution.LifecycleAwareValue;
+import org.gradle.internal.execution.Result;
+import org.gradle.internal.execution.Step;
+
+import java.util.ArrayList;
+import java.util.List;
+
+public class FinalizePropertiesStep<C extends Context, R extends Result> implements Step<C, R> {
+    private final Step<? super C, ? extends R> delegate;
+
+    public FinalizePropertiesStep(Step<? super C, ? extends R> delegate) {
+        this.delegate = delegate;
+    }
+
+    @Override
+    public R execute(C context) {
+        List<LifecycleAwareValue> lifecycleAwareValues = new ArrayList<>();
+        context.getWork().visitInputFileProperties((propertyName, value, incremental, fingerprinter) -> {
+            if (value instanceof LifecycleAwareValue) {
+                LifecycleAwareValue lifecycleAwareValue = (LifecycleAwareValue) value;
+                lifecycleAwareValues.add(lifecycleAwareValue);
+                lifecycleAwareValue.prepareValue();
+            }
+        });
+        try {
+            return delegate.execute(context);
+        } finally {
+            lifecycleAwareValues.forEach(LifecycleAwareValue::cleanupValue);
+        }
+    }
+}
